UPDATE: Correciones de metodos de publicacion y service
- Se crean dtos para crear resenas y publicaciones
- Se agregan validaciones a los dtos
- Se borran metodos innecesarios
- Se borran atributos innecesarios

+++ src/main/java/com/example/LaFachada/Dto/PublicacionDTO.java
@@ -26,7 +26,7 @@ public class PublicacionDTO {
 
     @NotBlank(message = "El precio es obligatorio")
     @Size(min = 1, message = "El precio debe ser un número positivo")
-    private Integer precio;
+    private Double precio;
 
 
     @NotBlank(message = "La ubicación es obligatoria")
